// src/api/stetsom/model/productBlock.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Stetsom API
 * API documentation for Stetsom
 * OpenAPI spec version: 1.0.0
 */
import type { ProductBlockData } from "./productBlockData";
import type { ProductBlockType } from "./productBlockType";

export interface ProductBlock {
  block_id: string;
  type: ProductBlockType;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  order: number;
  data: ProductBlockData;
}